fix bugs and update ui for date select

+++ src/main/java/com/celivra/personnalms/DeletePeople.java
@@ -5,6 +5,7 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 import java.io.IOException;
 
@@ -12,7 +13,13 @@ public class DeletePeople extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String id = req.getParameter("deleteID");
-        PeopleService.removeById(id);
+        HttpSession session = req.getSession();
+        if(PeopleService.findById(id) == null) {
+           session.setAttribute("DeleteError", "Can't delete people");
+        }else{
+            PeopleService.removeById(id);
+            session.setAttribute("DeleteComplete", "delete people successfully");
+        }
         resp.sendRedirect("/Dashboard.jsp");
     }
 }
